Separate ResultHandler.h and -inl.h
