+++ CMakeLists.txt
@@ -26,6 +26,8 @@ set(LIBRARY_SOURCE_FILES
     include/eventkit/promise/Resolver.h
     include/eventkit/promise/ResultHandler.h
     include/eventkit/promise/detail/ResultHandler-inl.h
+    include/eventkit/promise/detail/FunctionResultHandler.h
+    include/eventkit/promise/detail/FunctionResultHandler-inl.h
     include/eventkit/promise/detail/PromiseCore.h
     include/eventkit/promise/detail/ResultHandlerMultipleInheritanceHelper.h
     include/eventkit/promise/detail/ThenTransformationCore.h
